fix VersionChecker for QUnit and enable test
